Add levels 3-5
Level 3 - add field star to Post entity and extra methods for it
Level 4 - add entity Comment, classes and methods for work with comments
Level 5 - add method to PostController that returns post with all nested comments in it
Add DTO classes
Add tests for PostController and CommentController
Refactoring code adding Optional<>

+++ src/main/java/com/lytvynova/service/PostServiceImpl.java
@@ -1,5 +1,6 @@
 package com.lytvynova.service;
 
+import com.lytvynova.dto.PostWithoutCommentDto;
 import com.lytvynova.entity.Post;
 import com.lytvynova.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -21,8 +23,18 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public List<Post> findAllPostsByTitle(String title) {
-        return postRepository.findAllByTitleIgnoreCase(title);
+        List<Post> posts = postRepository.findAllByTitleIgnoreCase(title);
+
+        return posts;
     }
+
+    @Override
+    public List<Post> findAllTopPosts() {
+        List<Post> posts = postRepository.findByStarTrue();
+
+        return posts;
+    }
+
 
     @Override
     public Post addPost(Post post) {
@@ -31,21 +43,59 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public Post editPost(Long id, Post post) {
-        Post postToUpdate = postRepository.findById(id).get();
+        Optional<Post> postOptional = postRepository.findById(id);
+        if (postOptional.isEmpty()) {
+            throw new IllegalArgumentException("there is no post with provided id");
+        }
+        Post postToUpdate = postOptional.get();
+        postToUpdate.setTitle(post.getTitle());
+        postToUpdate.setContent(post.getContent());
 
-        if (Objects.nonNull(post.getTitle()) &&
-                !"".equalsIgnoreCase(post.getTitle())) {
-            postToUpdate.setTitle(post.getTitle());
-        }
-        if (Objects.nonNull(post.getContent()) &&
-                !"".equalsIgnoreCase(post.getContent())) {
-            postToUpdate.setContent(post.getContent());
-        }
+
         return postRepository.save(postToUpdate);
     }
 
     @Override
+    public Post markTopPost(Long id) {
+        Optional<Post> postOptional = postRepository.findById(id);
+        if (postOptional.isEmpty()) {
+            throw new IllegalArgumentException("there is no post with provided id");
+        }
+        Post postToMarkAsTop = postOptional.get();
+        postToMarkAsTop.setStar(true);
+       return postRepository.save(postToMarkAsTop);
+
+    }
+
+    @Override
+    public Post deleteTopMark(Long id) {
+        Optional<Post> postOptional = postRepository.findById(id);
+        if (postOptional.isEmpty()) {
+            throw new IllegalArgumentException("there is no post with provided id");
+        }
+        Post postToDeleteMark = postOptional.get();
+        postToDeleteMark.setStar(false);
+
+        return postRepository.save(postToDeleteMark);
+    }
+
+    @Override
+    public Post getPostWithComments(Long id) {
+        Optional<Post> postOptional = postRepository.findById(id);
+        if (postOptional.isEmpty()) {
+            throw new IllegalArgumentException("there is no post with provided id");
+        }
+        return postOptional.get();
+
+    }
+
+
+    @Override
     public void deletePost(Long id) {
+        Optional<Post> postOptional = postRepository.findById(id);
+        if (postOptional.isEmpty()) {
+            throw new IllegalArgumentException("there is no post with provided id");
+        }
         postRepository.deleteById(id);
     }
 
@@ -54,5 +104,15 @@ public class PostServiceImpl implements PostService {
         return postRepository.findAll(Sort.by(Sort.Direction.ASC, sortParameter));
     }
 
+    private PostWithoutCommentDto toPostWithoutCommentDto(Post post) {
+        PostWithoutCommentDto postWithoutCommentDto = new PostWithoutCommentDto();
 
+        postWithoutCommentDto.setId(post.getId());
+        postWithoutCommentDto.setTitle(post.getTitle());
+        postWithoutCommentDto.setContent(post.getContent());
+        postWithoutCommentDto.setStar(post.isStar());
+
+        return postWithoutCommentDto;
+
+    }
 }
